+++ pydirectus/session.py
@@ -1,7 +1,10 @@
 import logging
 from time import time
 import httpx
+from httpx import HTTPError
 from dataclasses import dataclass
+from tenacity import retry, stop_after_attempt, wait_exponential
+
 
 @dataclass
 class APIResponse:
@@ -13,7 +16,7 @@ class APIResponse:
 
 class Session():
     "handle lowlevel requests to the Directus API"
-    def __init__(self, url: str, token: str) -> None:
+    def __init__(self, url: str, token: str, timeout: float = 20.0) -> None:
 
         if not url:
             raise ValueError("URL not provided")
@@ -24,7 +27,9 @@ class Session():
 
         self.url = url
         self.token = token
+        self.client = httpx.Client(timeout=timeout)
 
+    @retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=4, max=10))
     def get(self, endpoint: str) -> APIResponse:
         "GET request"
 
@@ -34,26 +39,37 @@ class Session():
 
         # Make request
         start = time()
-        response = httpx.get(url, headers=headers)
+        response = self.client.get(url, headers=headers)
         duration = max(int(time() - start), 1) # avoid 0ms response
+
+        # retry logic
+        if not response.is_success:
+            logging.warning(f'request to {url} timed out - retrying 3 times')
+            raise HTTPError(f"{response.status_code} {response.text}")
 
         # parse response and returns it
         return self._make_response(url, duration, response)
 
+    @retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=4, max=10))
     def post(self, endpoint: str, data: dict) -> APIResponse:
         "POST request"
-
         # build URL and headers
         url = self._make_url(endpoint)
         headers = self._make_headers()
 
         # Make request
         start = time()
-        response = httpx.post(url, headers=headers, json=data)
+        response = self.client.post(url, headers=headers, json=data)
         duration = max(int(time() - start), 1)
+
+        # retry logic
+        if not response.is_success:
+            logging.warning(f'request to {url} timed out - retrying 3 times')
+            raise HTTPError(f"{response.status_code} {response.text}")
 
         return self._make_response(url, duration, response)
 
+    @retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=4, max=10))
     def patch(self, endpoint: str, data: dict) -> APIResponse:
         "PATCH request"
 
@@ -63,11 +79,17 @@ class Session():
 
         # Make request
         start = time()
-        response = httpx.patch(url, headers=headers, json=data)
+        response = self.client.patch(url, headers=headers, json=data)
         duration = max(int(time() - start), 1)
+
+        # retry logic
+        if not response.is_success:
+            logging.warning(f'request to {url} timed out - retrying 3 times')
+            raise HTTPError(f"{response.status_code} {response.text}")
 
         return self._make_response(url, duration, response)
 
+    @retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=4, max=10))
     def delete(self, endpoint: str, data: dict) -> APIResponse:
         "DELETE request"
 
@@ -78,13 +100,18 @@ class Session():
         # Make request
         start = time()
         # have to do custom to pass payload to the DELETE request...
-        with httpx.Client() as client:
-            response = client.request(method="DELETE", url=url,
-                                      headers=headers, json=data)
+        response = self.client.request(method="DELETE", url=url,
+                                       headers=headers, json=data)
         duration = max(int(time() - start), 1)
+
+        # retry logic
+        if not response.is_success:
+            logging.warning(f'request to {url} timed out - retrying 3 times')
+            raise HTTPError(f"{response.status_code} {response.text}")
 
         return self._make_response(url, duration, response)
 
+    @retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=4, max=10))
     def search(self, endpoint: str, data: dict) -> APIResponse:
         "Search request"
 
@@ -93,14 +120,19 @@ class Session():
         headers = self._make_headers()
         # Make request
         start = time()
-        with httpx.Client() as client:
-            response = client.request(method="SEARCH", url=url,
-                                      headers=headers, json=data)
+        response = self.client.request(method="SEARCH", url=url,
+                                       headers=headers, json=data)
         duration = max(int(time() - start), 1)
+
+        # retry logic
+        if not response.is_success:
+            logging.warning(f'request to {url} timed out - retrying 3 times')
+            raise HTTPError(f"{response.status_code} {response.text}")
 
         return self._make_response(url, duration, response)
 
     # files management
+    @retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=4, max=10))
     def download(self, endpoint: str) -> bytes:
         "Download a file"
 
@@ -109,13 +141,19 @@ class Session():
         headers = self._make_headers()
 
         # Make request
-        response = httpx.get(url, headers=headers)
+        response = self.client.get(url, headers=headers)
+
+        # retry logic
+        if not response.is_success:
+            logging.warning(f'request to {url} timed out - retrying 3 times')
+            raise HTTPError(f"{response.status_code} {response.text}")
 
         if response.status_code != 200:
             logging.error(f"{url}: error {response.status_code}: {response.text}")
             return b''
         return response.content
 
+    @retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=4, max=10))
     def upload(self, endpoint: str, data: dict, files: dict) -> APIResponse:
         "Upload a file"
         url = self._make_url(endpoint)
@@ -123,13 +161,19 @@ class Session():
         start = time()
         response = httpx.post(url, headers=headers, data=data, files=files)
         duration = max(int(time() - start), 1)
+
+        # retry logic
+        if not response.is_success:
+            logging.warning(f'request to {url} timed out - retrying 3 times')
+            raise HTTPError(f"{response.status_code} {response.text}")
+
         return self._make_response(url, duration, response)
 
 
 
     def ping(self) -> int:
         "Check if the API is reachable"
-        path = "server/health"
+        path = "server/ping"
         url = self._make_url(path)
         headers = self._make_headers()