+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `pydirectus` will be documented in this file.
 
+
+## [0.2.x] - 2024-09-22
+
+### Added
+ - Session retry support (3 attempts)
+ - httpx configurable timeout.
+
 ## [0.2.0] - 2024-09-21
 
 ### Added